Move Context C++ definitions to context.cc
