+++ v8go.cc
@@ -74,18 +74,6 @@ static RtnError ExceptionError(TryCatch& try_catch,
   return rtn;
 }
 
-m_value* tracked_value(m_ctx* ctx, m_value* val) {
-  // (rogchap) we track values against a context so that when the context is
-  // closed (either manually or GC'd by Go) we can also release all the
-  // values associated with the context;
-  if (val->id == 0) {
-    val->id = ++ctx->nextValId;
-    ctx->vals[val->id] = val;
-  }
-
-  return val;
-}
-
 m_unboundScript* tracked_unbound_script(m_ctx* ctx, m_unboundScript* us) {
   ctx->unboundScripts.push_back(us);
 
@@ -586,27 +574,6 @@ ContextPtr NewContext(IsolatePtr iso,
 
 int ContextRetainedValueCount(ContextPtr ctx) {
   return ctx->vals.size();
-}
-
-void ContextFree(ContextPtr ctx) {
-  if (ctx == nullptr) {
-    return;
-  }
-  ctx->ptr.Reset();
-
-  for (auto it = ctx->vals.begin(); it != ctx->vals.end(); ++it) {
-    auto value = it->second;
-    value->ptr.Reset();
-    delete value;
-  }
-  ctx->vals.clear();
-
-  for (m_unboundScript* us : ctx->unboundScripts) {
-    us->ptr.Reset();
-    delete us;
-  }
-
-  delete ctx;
 }
 
 RtnValue RunScript(ContextPtr ctx, const char* source, const char* origin) {
